% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GenerateASingleCandidate}
\alias{GenerateASingleCandidate}
\title{\code{GenerateASingleCandidate} Generates a novel sample from a target class and evaluate it against the other classes to check if it satisfies the confidence level
returns NA if the generated sample does not satisfy the condition, otherwise returns novel sample}
\usage{
GenerateASingleCandidate(data, Class, col, Prob)
}
\arguments{
\item{data}{A dataframe containing available data}

\item{Class}{The target class}

\item{col}{Column the target class is located in the dataframe}

\item{Prob}{Minimum confidence level to generate sample}
}
\value{
A dataframe containing a novel sample if it satisfies the confidence given in Prob, otherwise NA
}
\description{
\code{GenerateASingleCandidate} Generates a novel sample from a target class and evaluate it against the other classes to check if it satisfies the confidence level
returns NA if the generated sample does not satisfy the condition, otherwise returns novel sample
}
