% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GenerateMultipleCandidates}
\alias{GenerateMultipleCandidates}
\title{\code{GenerateMultipleCandidates} Asks for a dataframe and some parameters and returns multiple novel samples from the target class}
\usage{
GenerateMultipleCandidates(data, Class, col, Prob, amount)
}
\arguments{
\item{data}{Reference dataframe}

\item{Class}{Value of the target class}

\item{col}{column of the dataframe which contains the class}

\item{Prob}{Minimum confidence level to generate the sample}

\item{amount}{Number of novel samples to be generated}
}
\value{
A dataframe containing novel samples of the class \code{Class} that satisfies \code{Prob} confidence.
}
\description{
\code{GenerateMultipleCandidates} Asks for a dataframe and some parameters and returns multiple novel samples from the target class
}
\examples{
# basic usage of GenerateMultipleCandidates
GenerateMultipleCandidates(iris,Class='virginica',col=5,Prob=0.3,amount=10)
}
