% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ObtainCandidate}
\alias{ObtainCandidate}
\title{\code{ObtainCandidate} Asks for a vector and returns a value along with the range it is contained in the attribute
Is used alongside other functions when generating a new sample}
\usage{
ObtainCandidate(Dado)
}
\arguments{
\item{Dado}{vector containing an attribute of your dataframe}
}
\value{
Obtains a sample of an attribute based on their possible values, it is a part of the \code{Generate} function
}
\description{
\code{ObtainCandidate} Asks for a vector and returns a value along with the range it is contained in the attribute
Is used alongside other functions when generating a new sample
}
