% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Generate}
\alias{Generate}
\title{\code{Generate} Asks for a dataframe and generates a new sample
returns novel sample along with intervals it contained to revalidate it using confidence levels}
\usage{
Generate(data, regression = FALSE)
}
\arguments{
\item{data}{Dataframe}

\item{regression}{if we are to generate data for regression or classification (will the data be conditioned on a certain class)}
}
\description{
\code{Generate} Asks for a dataframe and generates a new sample
returns novel sample along with intervals it contained to revalidate it using confidence levels
}
